+++ boards/board.h
@@ -12,6 +12,8 @@
 
 #if defined(TESSERAE_BOARD_WAVESHARE_133E6)
 #  include "waveshare_133e6.h"
+#elif defined(TESSERAE_BOARD_WAVESHARE_1085G)
+#  include "waveshare_1085g.h"
 #elif defined(TESSERAE_BOARD_SEEED_E1004)
 #  include "seeed_reterminal_e1004.h"
 #elif defined(TESSERAE_BOARD_SEEED_E1002)